Update configure-rootfs anchor for typo.
This part

> Consider skimming the rootfs section and the global section of the installation docs to understand the significance of this (then return here).

was referencing to `#configure-rootfs` already. A quick `ag` on the repo did not find any other internal reference to the `#configure-roofs` typo, but **this might break external links**. In the event the external links are more important than correctness, the fix should instead change that link to use the erroneous `#configure-roofs`.